+++ packages/java/endpoint/src/main/java/dev/hilla/crud/CrudService.java
@@ -6,7 +6,7 @@ import dev.hilla.Nullable;
  * A browser-callable service that can create, read and update a given type of
  * object.
  */
-public interface CrudService<T, ID> extends ListService<T> {
+public interface CrudService<T, ID> extends ListService<T>, GetService<T, ID> {
 
     /**
      * Saves the given object and returns the (potentially) updated object.
@@ -20,26 +20,6 @@ public interface CrudService<T, ID> extends ListService<T> {
      */
     @Nullable
     T save(T value);
-
-    /**
-     * Gets the object with the given id.
-     *
-     * @param id
-     *            the id of the object
-     * @return the object, or null if no object was found
-     */
-    T get(ID id);
-
-    /**
-     * Checks if an object with the given id exists.
-     *
-     * @param id
-     *            the id of the object
-     * @return {@code true} if the object exists, {@code false} otherwise
-     */
-    default boolean exists(ID id) {
-        return get(id) != null;
-    }
 
     /**
      * Deletes the object with the given id.
